Refactor: Stronger separate parsing and translation steps
